+++ src/Repository/Model/Artist.php
@@ -1025,7 +1025,7 @@ class Artist extends database_object implements library_item, GarbageCollectible
                 }
                 $updated    = true;
                 $current_id = $artist_id;
-                //Stats::migrate('artist', $this->id, $artist_id);
+                Stats::migrate('artist', $this->id, $artist_id, $song_id);
                 Useractivity::migrate('artist', $this->id, $artist_id);
                 Recommendation::migrate('artist', $this->id);
                 Share::migrate('artist', $this->id, $artist_id);
@@ -1153,8 +1153,8 @@ class Artist extends database_object implements library_item, GarbageCollectible
             // fill missing artist_map rows
             $sql = "INSERT IGNORE INTO `artist_map` (`artist_id`, `object_type`, `object_id`) SELECT DISTINCT `song`.`artist` AS `artist_id`, 'song', `song`.`id` FROM `song` WHERE `song`.`artist` = ? UNION SELECT DISTINCT `album`.`album_artist` AS `artist_id`, 'album', `album`.`id` FROM `album` WHERE `album`.`album_artist` = ?;";
             Dba::write($sql, array($artist_id, $artist_id));
-            // fill missing album_map rows
-            $sql = "INSERT IGNORE INTO `album_map` (`album_id`, `object_type`, `object_id`) SELECT DISTINCT `artist_map`.`object_id` AS `album_id`, 'album' AS `object_type`, `artist_map`.`artist_id` AS `object_id` FROM `artist_map` WHERE `artist_map`.`object_type` = 'album' AND `artist_map`.`object_id` IS NOT NULL AND `artist_map`.`artist_id` = ? UNION SELECT DISTINCT `song`.`album` AS `album_id`, 'song' AS `object_type`, `song`.`artist` AS `object_id` FROM `song` WHERE `song`.`album` IS NOT NULL AND `song`.`artist` = ? UNION SELECT DISTINCT `song`.`album` AS `album_id`, 'song' AS `object_type`, `artist_map`.`artist_id` AS `object_id` FROM `artist_map` LEFT JOIN `song` ON `artist_map`.`object_type` = 'song' AND `artist_map`.`object_id` = `song`.`id` WHERE `song`.`album` IS NOT NULL AND `artist_map`.`object_type` = 'song' AND `artist_map`.`artist_id` = ?;";
+            // fill missing albumartist_map rows
+            $sql = "INSERT IGNORE INTO `albumartist_map` (`album_id`, `object_type`, `object_id`) SELECT DISTINCT `artist_map`.`object_id` AS `album_id`, 'album' AS `object_type`, `artist_map`.`artist_id` AS `object_id` FROM `artist_map` WHERE `artist_map`.`object_type` = 'album' AND `artist_map`.`object_id` IS NOT NULL AND `artist_map`.`artist_id` = ? UNION SELECT DISTINCT `song`.`album` AS `album_id`, 'song' AS `object_type`, `song`.`artist` AS `object_id` FROM `song` WHERE `song`.`album` IS NOT NULL AND `song`.`artist` = ? UNION SELECT DISTINCT `song`.`album` AS `album_id`, 'song' AS `object_type`, `artist_map`.`artist_id` AS `object_id` FROM `artist_map` LEFT JOIN `song` ON `artist_map`.`object_type` = 'song' AND `artist_map`.`object_id` = `song`.`id` WHERE `song`.`album` IS NOT NULL AND `artist_map`.`object_type` = 'song' AND `artist_map`.`artist_id` = ?;";
             Dba::write($sql, array($artist_id, $artist_id, $artist_id));
             $params = array($artist_id);
             // artist.time